+++ src/controllers/api/users/index.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import * as users from "./UuserHandlers";
+import * as users from "./user_handlers";
 
 // /api/users
 const router = Router();
